refactor(mention): extract caret helper to lib/textarea-caret
Move the textarea caret-measuring helper (getCaretRect + the copied-style
list) out of the component into a new registry:lib item so base.tsx stays
focused on the mention behaviour. It's an inlined, dependency-free equivalent
of the `textarea-caret` npm package the React Aria example uses.

mention now depends on `textarea-caret` (registryDependencies) and imports
`getCaretRect` from `@/registry/lib/textarea-caret`.
